docs(getting-started): Merged pretty print with start.py
